Encode and decode data going to and coming from sockets
Under python3, we are responsible for decoding and encoding. The
implementation does not change the behaviour for python 2, and by
default uses python2-esque behaviour (using sys.getdefaultencoding())
under python 3, but lets the user override the encoding or even set it
to None to get bytes objects back. It also allows the user to pass bytes
objects to put, striking the best balance between backwards
compatibility for binary data in tubes (which can never be achieved
100%) and working nicely with text data.

Tested with python 2.4, 2.7 and 3.4